pyproject.toml
[project]
name = "tweet-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.6",
    "langgraph>=0.5.4",
    "python-dotenv>=1.1.1",
    "tweepy>=4.16.0",
    "requests>=2.31.0",
]
